+++ src/components/AddForm.js
@@ -1,10 +1,8 @@
 import "../styles/AddForm.css";
-import FormInput from "./FormInput";
-import UploadPhotos from "./UploadPhotos";
 import { useState } from "react";
 import { addTripsToLocalStorage } from "../services/tripsStorage";
-import DatePicker from "react-multi-date-picker";
 import { useHistory, Link } from "react-router-dom";
+import Form from "./Form";
 
 function AddForm() {
   const [destinationInput, setDestinationInput] = useState("");
@@ -60,78 +58,22 @@ function AddForm() {
 
   return (
     <div className="form">
-      <form className="form__textfields" onSubmit={handleSubmit}>
-        <DatePicker
-          value={datesInput}
-          onChange={(date) => setDatesInput(date)}
-          format="MM/DD/YYYY"
-          placeholder="select travel dates"
-          range
-          inputClass="custom-input"
-          required
-        />
-        <FormInput
-          id="destination"
-          name="destination"
-          value={destinationInput}
-          onChange={(e) => {
-            setDestinationInput(e.target.value);
-          }}
-          // required="true"
-        />
-        <FormInput
-          id="activities"
-          name="activities"
-          value={activitiesInput}
-          onChange={(e) => {
-            setActivitiesInput(e.target.value);
-          }}
-        />
-        <FormInput
-          id="locations"
-          name="locations"
-          value={locationsInput}
-          onChange={(e) => {
-            setLocationsInput(e.target.value);
-          }}
-        />
-        <FormInput
-          id="notes"
-          name="notes"
-          value={notesInput}
-          onChange={(e) => {
-            setNotesInput(e.target.value);
-          }}
-        />
-        <UploadPhotos
-          id="photo"
-          name="photo"
-          onChange={(e) => {
-            setImageUploads(e.target.files);
-
-            const imageArray = Array.from(e.target.files).map((file) =>
-              URL.createObjectURL(file)
-            );
-            setImgPreview([]);
-            setImgPreview((prevURL) => prevURL.concat(imageArray));
-          }}
-        />
-
-        {imgPreview
-          ? imgPreview.map((imgPreview) => {
-              return (
-                <img className="imagePreview" src={imgPreview} alt="preview" />
-              );
-            })
-          : null}
-
-        <div className="form__buttons">
-          <button type="submit" className="submit">
-            save
-          </button>
-          <button className="cancel">cancel</button>
-        </div>
-      </form>
+      <Form
+        onSubmit={handleSubmit}
+        destinationInput={destinationInput}
+        setDestinationInput={setDestinationInput}
+        datesInput={datesInput}
+        setDatesInput={setDatesInput}
+        activitiesInput={activitiesInput}
+        setActivitiesInput={setActivitiesInput}
+        locationsInput={locationsInput}
+        setLocationsInput={setLocationsInput}
+        notesInput={notesInput}
+        setNotesInput={setNotesInput}
+        imgPreview={imgPreview}
+        setImgPreview={setImgPreview}
+        setImageUploads={setImageUploads}
+      />
       <Link to="/destinationShuffle">
         <p className="needInspo">need some inspiration?</p>
       </Link>
