+++ src/components/Trips.js
@@ -27,10 +27,9 @@ function Trips() {
   function renderTrips() {
     return trips.map((trips, index) => {
       return (
-        <div>
+        <div key={index}>
           <TripCard
             path={`/saved-trip/${trips.id}`}
-            key={index}
             src={`https://source.unsplash.com/random/?${trips.destination}`}
             alt={trips.destination}
             text={trips.destination}
